+++ example-app/app/core/containers/view/view.component.ts
@@ -7,13 +7,16 @@ import {
 import { ActivatedRoute } from '@angular/router';
 
 import { Observable, Subject } from 'rxjs';
-import { slice, Slice } from 'lib';
+import { Slice, Sliced } from 'lib';
 import {
   APP_HUB_KEY,
   AppChildNotes,
   AppNotes
-} from '../../../routing/hub/app.hub';
-import { BOLID_HUB_KEY, BolidNotes } from '../../../views/bolid/hub/bolid.hub';
+} from '../../../routing/hub/app.notes';
+import {
+  BOLID_HUB_KEY,
+  BolidNotes
+} from '../../../views/bolid/hub/bolid.notes';
 
 @Component({
   selector: 'app-view',
@@ -25,10 +28,10 @@ export class ViewComponent implements OnInit, OnDestroy {
   public vehicles: Observable<string[]>;
   private unsubscribe$: Subject<boolean>;
 
-  @slice(APP_HUB_KEY)
+  @Sliced(APP_HUB_KEY)
   private app: Slice<AppNotes, AppChildNotes>;
 
-  @slice(BOLID_HUB_KEY)
+  @Sliced(BOLID_HUB_KEY)
   private bolid: Slice<BolidNotes>;
 
   constructor(private readonly route: ActivatedRoute) {}
