+++ example-app/app/views/automobile/hub/automobile.notes.ts
