+++ example-app/app/views/bike/hub/bike.notes.ts
