+++ example-app/app/views/bike/hub/bike.routes.ts
@@ -4,7 +4,7 @@ import { createFeature, Slice } from 'lib';
 import { BikeComponent } from '../container/bike.component';
 import { VehiclesResolver } from '../../../routing/resolvers/vehicles.resolver';
 import { appSlice } from '../../../routing/hub/app.routes';
-import { BIKE_HUB_KEY, BikeNotes } from './bike.hub';
+import { BIKE_HUB_KEY, BikeNotes } from './bike.notes';
 
 export const routes = [
   {
