+++ example-app/app/views/bolid/hub/bolid.notes.ts
